# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления рекламодателями и их рекламными кампаниями.",
        "title": "Eshke Reklama API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/api/ad/request": {
            "post": {
                "description": "Возвращает объявление для вставленного на сайт партнёра рекламного блока по embed_token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public_ads"
                ],
                "summary": "Запрос рекламы для публичного блока",
                "parameters": [
                    {
                        "description": "Публичный токен рекламного блока",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AdRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AdRequestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    }
                }
            }
        },
...
        "/api/ad_campaigns/{ad_campaign_id}/ad_groups": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ad_groups"
                ],
                "summary": "Список групп объявлений",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID рекламной кампании",
                        "name": "ad_campaign_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ListAdGroupsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ad_groups"
                ],
                "summary": "Создание группы объявлений",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID рекламной кампании",
                        "name": "ad_campaign_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Параметры группы",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateAdGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateAdGroupResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    }
                }
            }
        },
        "/api/ad_campaigns/{ad_campaign_id}/ad_groups/{ad_group_id}": {
            "put": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ad_groups"
                ],
                "summary": "Обновление группы объявлений",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID рекламной кампании",
                        "name": "ad_campaign_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID группы",
                        "name": "ad_group_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Поля для обновления",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateAdGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpx.Success"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ad_groups"
                ],
                "summary": "Удаление группы объявлений",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID рекламной кампании",
                        "name": "ad_campaign_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID группы",
                        "name": "ad_group_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpx.Success"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AdGroupResponse": {
            "type": "object",
            "properties": {
                "age_from": {
                    "type": "integer"
                },
                "age_to": {
                    "type": "integer"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "topic": {
                    "type": "string"
                }
            }
        },
        "dto.AdRequest": {
            "type": "object",
            "required": [
                "embed_token"
            ],
            "properties": {
                "embed_token": {
                    "type": "string"
                },
                "visitor_id": {
                    "type": "string"
                }
            }
        },
        "dto.AdRequestResponse": {
            "type": "object",
            "properties": {
                "ad": {
                    "$ref": "#/definitions/dto.AdResponse"
                },
                "click_url": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "dto.AdResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "short_desc": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "target_url": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "dto.CreateAdGroupRequest": {
            "type": "object",
            "required": [
                "age_from",
                "age_to",
                "gender",
                "name",
                "region",
                "topic"
            ],
            "properties": {
                "age_from": {
                    "type": "integer"
                },
                "age_to": {
                    "type": "integer"
                },
                "gender": {
                    "type": "string",
                    "enum": [
                        "man",
                        "woman",
                        "any"
                    ]
                },
                "name": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "topic": {
                    "type": "string"
                }
            }
        },
        "dto.CreateAdGroupResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.ListAdGroupsResponse": {
            "type": "object",
            "properties": {
                "ad_campaign_id": {
                    "type": "integer"
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.AdGroupResponse"
                    }
                }
            }
        },
...
        "dto.UpdateAdGroupRequest": {
            "type": "object",
            "properties": {
                "age_from": {
                    "type": "integer",
                    "minimum": 1
                },
                "age_to": {
                    "type": "integer",
                    "minimum": 1
                },
                "gender": {
                    "type": "string",
                    "minLength": 1,
                    "enum": [
                        "man",
                        "woman",
                        "any"
                    ]
                },
                "name": {
                    "type": "string",
                    "minLength": 1
                },
                "region": {
                    "type": "string",
                    "minLength": 1
                },
                "topic": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
...
        "httpx.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "httpx.Success": {
            "type": "object",
            "properties": {
                "data": {}
            }
        }
    },
    "securityDefinitions": {
        "CookieAuth": {
            "type": "apiKey",
            "name": "session_id",
            "in": "cookie"
        }
    }
}